Adding Firebase persistance for #8.
Moving to an opt-in model, there won't be a reaction unless user has explicitly asked to be informed of offending words.

Sadly, Firebase JS SDK does not run in strict mode in Node. Reported and waiting for response. In the meantime we need to add "use strict" at the top of each source file : (

TODO: command to be interested.